# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gym de los pibes",
    "version": "1.0.0",
    "description": "Documentación de la API del gym"
  },
  "paths": {
...
    "/actividades/{id}": {
      "get": {
        "summary": "Obtener una actividad por ID",
        "operationId": "getActividadById",
        "tags": ["Actividades"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Actividad encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actividad"
                }
              }
            }
          },
          "404": {
            "description": "Actividad no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Actualizar una actividad existente (Requiere token de Admin)",
        "operationId": "updateActividad",
        "tags": ["Actividades"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActividadDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actividad actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actividad"
                }
              }
            }
          },
          "400": {
            "description": "Error en la solicitud",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
             "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prohibido (Sin permisos de Admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Actividad no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Eliminar una actividad (Requiere token de Admin)",
        "operationId": "deleteActividad",
        "tags": ["Actividades"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Actividad eliminada"
          },
           "401": {
            "description": "No autorizado",
             "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prohibido (Sin permisos de Admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Actividad no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/inscripciones/all": {
      "get": {
        "summary": "Listar todas las inscripciones (Requiere token de Admin)",
        "operationId": "getAllInscripciones",
        "tags": ["Inscripciones"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de todas las inscripciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Inscripcion"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prohibido (Requiere rol de Administrador)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/inscripciones/{inscripcionId}": {
      "get": {
        "summary": "Obtener una inscripción por su ID. Requiere ser el usuario de la inscripción o un administrador.",
        "operationId": "getInscripcionById",
        "tags": ["Inscripciones"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "inscripcionId",
            "in": "path",
            "required": true,
            "description": "ID de la inscripción a obtener",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la inscripción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inscripcion"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prohibido (No es el usuario de la inscripción ni un administrador)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Inscripción no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Actividad": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "titulo": { "type": "string" },
          "descripcion": { "type": "string" },
          "dia": { "type": "string" },
          "horario": { "type": "string", "format": "time", "example": "10:00:00" },
          "duracion": { "type": "integer", "description": "Duración en minutos" },
          "cupo": { "type": "integer" },
          "categoria": { "type": "string" },
          "imagen_url": { "type": "string", "format": "url" },
          "active": { "type": "boolean", "default": true }
        }
      },
      "ActividadDTO": {
        "type": "object",
        "required": ["titulo", "dia", "horario", "duracion", "cupo", "categoria"],
        "properties": {
          "titulo": { "type": "string" },
          "descripcion": { "type": "string" },
          "dia": { "type": "string" },
          "horario": { "type": "string", "format": "time", "example": "10:00:00" },
          "duracion": { "type": "integer", "description": "Duración en minutos" },
          "cupo": { "type": "integer" },
          "categoria": { "type": "string" },
          "imagen_url": { "type": "string", "format": "url" }
        }
      },
...
      "Inscripcion": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID de la inscripción"},
          "usuarioId": { "type": "integer", "format": "int64" },
          "actividadId": { "type": "integer", "format": "int64" },
          "fechaInscripcion": { "type": "string", "format": "date-time" },
          "actividad": { "$ref": "#/components/schemas/Actividad" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "description": "Mensaje descriptivo del error" },
          "code": { "type": "string", "nullable": true, "description": "Código de error interno (opcional)" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}